simx86: return back nodes GC removed in 800e2169
We should be able to clean pages on node insertion, if
we do that _after_ the node is inserted, rather than before,
as it was prior to 800e2169 .
If we do that before, then the page may get unprotected and
removed from cache. And it may appear to be the same page that
we insert the node for, which will lead to bad things.

So add the GC after an insertion, and make sure it doesn't run
in prejit mode. Background threads should not unprotect anything
to not race with main thread.

See #2634