simx86: fix memory corruption avltr_delete
Incorrect memory was freed resulting in bogus keys.

Makes the logic like it is in the original avltr_delete in libavl,
by not copying data but the pointer when data is moved from a deleted
tree node to one that is kept but for which the data is deleted.
As an added bonus this way the actual data (TNode) doesn't need
altering any more, no bkptr adjustement needed anymore.

Another regression from 65db4096c, seen with aggressive garbage
collect (100 nodes max).