trees: chg find-less to crossing-or-in-range [#2638]
... in InvalidateNodeRange().

Current technique has a few inoccuracies:
- As soon as it finds the equal-or-greater node, it was ending the
  search, returning the previously visited node. It didn't consider
  the fact that there may be more less-than nodes if we descend
  leftwards from the found equal-or-greater node.
- If the greater-or-equal node is actually greater and the previous
  is less (and does not overlap), then we can end up the search
  immediately, rather than proceeding to invalidation. Since there
  is nothing to invalidate in this case.

This patch also changes the lookup condition to crossing-or-in-range,
because if the lesser node doesn't overlap with our range at all,
then we don't need it during the invalidation step.

Its the hardest part of the change: debugging those "dead nodes"
is not fun. :(