simx86: improve NodeLinker during exec, eliminate global LastXNode
The NodeLinker had trouble with LastXNode being potentially deleted, but
it was also missing a lot of cases with the login in DoExec_Fast.

Just like with described here for QEmu:
https://www.qemu.org/docs/master/devel/tcg.html#direct-block-chaining

"Note that, on step 3 (tcg_gen_exit_tb()), in addition to the jump slot index,
the address of the TB just executed is also returned. This address corresponds
to the TB that will be patched; it may be different than the one that was
directly executed from the main loop if the latter had already been chained to
other TBs."

we needed to return the address of the TB that was just executed; the simplest
was to have ecx in JIT return the PC (key) of that TB in seqbase; then FindTree can
find the node, and avoid returning a dead node at the same time, so we
know for sure we can link it with simpler logic.

Unlinkable exits have ePC==seqbase so we can skip the FindTree for those.
In a FreeCOM compilation, by having many more nodes linked, the number of
FindTree calls dropped from 1.9B to 548M, giving significant speed improvements
(~30s on 3 minutes).