simx86: lret etc: new eip value in protmode via Sim_helper return
We can avoid setting TheCPU.eip if Sim_helper returns the new
eip in eax, and as vm86 iret already calls Sim_helper anyway,
use it for the cs:eip handling as well.

So now the only places left that set TheCPU.eip are RETl
and RETlisp in vm86, which are fully compiled there. Modifying
O_POP2 and A_SR_SH4 to avoid %eax is a little tricky.