simx86: for "push segreg", always use 16-bit moves
But adjust ESP by 4 in 32-bit mode.

This matches real CPUs:
https://www.felixcloutier.com/x86/push
"if the operand size is 32-bits, either a zero-extended value
is pushed on the stack or the segment selector is written
on the stack using a 16-bit move. For the last case, all
recent Intel Core and Intel Atom processors perform a 16-bit
move, leaving the upper portion of the stack location
unmodified."