simx86: DoExec/NodeLinker cleanups for F_SLFJ/MTRAP/loop
Using TheCPU.eip for the block we can reliably detect forever loops
for the last block. The node linker doesn't need to care about
blocks with F_SLFJ then any more. On the other hand it should
never link blocks that set trap exceptions (mode & MTRAP).

Flagging back refs in the node linker can then be limited to blocks
with F_FPOP (the original purpose of that flagging).

This all simplifies the fast loop in DoExec, it's only skipped
for one-offs (F_SPEC/F_LEAV) or if we want debug logging.