simx86: nodelinker: remove unlinked_jmp_targets field
It's redundant and was only useful for optimization in the fast
exec loop, but since we now check for `*pLastXKey != G->key`
in the loop we don't need it any more.

Instead always check for `link` to see if the node is linkable,
and then `ref` if it is linked.