simx86: eliminate O_POP, use POP1/2/3, add MOPT where possible
Convert remaining users to POP1/POP2/POP3, except for POPsp
which uses just POP1/POP2 without POP3.

Note: we need to be careful if something between O_POP2 and O_POP3
may modify %ecx in JIT, which contains the SS base. If so, don't
optimize, but force a reload if O_POP1 is next.