simx86: generate a linked jump instead of tail code in Close if possible
Almost always the tail code is put because of pre-existing code at
PC, detected by e_querymark. If that code starts exactly there and has
compatible mode and cs we can generate a jump to it and link it
straight away.

I observed a 3-4 second reduction (1m57s -> 1m54s) on a FreeCOM
compilation.