simx86: stop using e_querymark to stop codegen (except prejit)
Only stop if
1. there's a jump op generated, or
2. we're in single step mode (SINGLEBLOCK or TF), or
3. we encounter a compatible block we can jump into
other code is going to be invalidated in DoClose(), we're just
going to bulldoze over it here.

SPEC_PREJIT should avoid overwriting, but instead of using a gap,
I let it give up, by not calling Close() and resetting CurrIMeta,
thereby throwing away the prejitted code, instead of generating
an early tailcode.