simx86: only use TheCPU.key for first PC of block, delay break last ins
The dual use to signal non-linkable exits was confusing for BreakNode,
which relies on TheCPU.key to always be the first PC of the block.

We can instead check in NodeLinker if the node is linkable. Note that
already linked jumps can exit because of signals, in which case L->ref
is already set, so this is no longer fatal, but became a check that it
doesn't need linking again.

The reversal of part of dd01a3594 in d08026d was correct in that it can't
break a node after the last instruction, but
(1) with TheCPU.key being potentially wrong ExecOne could delete the
    wrong node, and
(2) to execute the last instruction we still need to delay free()ing that
    code

Tested with zx_emul.