simx86: replace EXCP_BREAKNODE with proper API [#2708]
We just need a different function for cpatch,
`int TryInvalidateNodeRange()`
that returns -1 if it failed to invalidate either node due to
self-hit.
This kinda re-introduces the dead nodes in tree, but with the
following very important difference: unlike in prev impls, such
dead nodes are now still marked as code-nodes and protected, so there
is no inconsistency between the tree content, protection and bitmaps.
They are getting removed on the next interp cycle.