simx86: unlink forward refs changing code (remove direct jumps)
So far NodeUnlinker only undid direct jumps for references *to* the node
that was removed, but for forward references *from* the node it wasn't
needed since the node would be deleted anyway.

However with BreakNodeHook excluding the last instruction the linking
jump instruction could still be kept and jump into a node that was also
deleted. We must therefore also undo the jump for forward references!

Fixes #2708